+++ Bonxai_ros/bonxai_rviz_plugins/src/bonxai_grid_display.cpp
@@ -14,12 +14,29 @@
 #include <string>
 #include <sstream>
 
+#include <QObject>  // NOLINT
+
+#include "rviz_common/visualization_manager.hpp"
+#include "rviz_common/properties/int_property.hpp"
+#include "rviz_common/properties/enum_property.hpp"
+#include "rviz_common/properties/float_property.hpp"
+#include "rviz_common/properties/status_property.hpp"
+
 #include "bonxai/bonxai.hpp"
+#include "bonxai_msgs/msg/bonxai.hpp"
 #include "bonxai_msgs/conversion.hpp"
 #include "bonxai_rviz_plugins/bonxai_grid_display.hpp"
 
 namespace bonxai_rviz_plugins
 {
+using rviz_common::properties::StatusProperty;
+
+enum BonxaiVoxelColorMode
+{
+  BONXAI_CELL_COLOR,
+  BONXAI_Z_AXIS_COLOR,
+  BONXAI_PROBABILITY_COLOR,
+};
 
 BonxaiGridDisplay::BonxaiGridDisplay()
 {
@@ -40,14 +57,22 @@ BonxaiGridDisplay::BonxaiGridDisplay()
   alpha_property_->setMin(0.0);
   alpha_property_->setMax(1.0);
 
+  scalar_threshold_property_ =
+      new rviz_common::properties::FloatProperty("Threshold",
+                                                 0.5,
+                                                 "Set scalar/probability threshold. "
+                                                 "Voxels with lower values are "
+                                                 "ignored.",
+                                                 this,
+                                                 SLOT(updateScalarThreshold()));
+
   // credit to: https://github.com/OctoMap/octomap_rviz_plugins/pull/20
   style_property_ = new rviz_common::properties::EnumProperty("Style",
                                                               "Boxes",
                                                               "Rendering mode to "
                                                               "use, in order of "
                                                               "computational "
-                                                              "complexity. Note that the 'Points' "
-                                                              "option is not scaled to voxel size.",
+                                                              "complexity.",
                                                               this,
                                                               SLOT(updateStyle()));
   style_property_->addOption("Points", rviz_rendering::PointCloud::RM_POINTS);
@@ -79,6 +104,60 @@ void BonxaiGridDisplay::unsubscribe()
   MFDClass::unsubscribe();
 }
 
+// method taken from octomap_server package
+void BonxaiGridDisplay::setColorFromZAxis(double z_pos,
+                                          double color_factor,
+                                          rviz_rendering::PointCloud::Point& point)
+{
+  int i{};
+  double m{};
+  double n{};
+  double f{};
+
+  double s = 1.0;
+  double v = 1.0;
+  double h =
+      (1.0 - std::min(std::max((z_pos - min_z_) / (max_z_ - min_z_), 0.0), 1.0)) *
+      color_factor;
+
+  h -= floor(h);
+  h *= 6;
+  i = floor(h);
+  f = h - i;
+  if (!(i & 1))
+  {
+    f = 1 - f;  // if i is even
+  }
+  m = v * (1 - s);
+  n = v * (1 - s * f);
+
+  switch (i)
+  {
+    case 6:
+    case 0:
+      point.setColor(v, n, m);
+      break;
+    case 1:
+      point.setColor(n, v, m);
+      break;
+    case 2:
+      point.setColor(m, v, n);
+      break;
+    case 3:
+      point.setColor(m, n, v);
+      break;
+    case 4:
+      point.setColor(n, m, v);
+      break;
+    case 5:
+      point.setColor(v, m, n);
+      break;
+    default:
+      point.setColor(1, 0.5, 0.5);
+      break;
+  }
+}
+
 void BonxaiGridDisplay::updateBonxaiColorMode()
 {
   MFDClass::updateTopic();
@@ -87,6 +166,16 @@ void BonxaiGridDisplay::updateBonxaiColorMode()
 void BonxaiGridDisplay::updateAlpha()
 {
   MFDClass::updateTopic();
+}
+
+void BonxaiGridDisplay::updateScalarThreshold()
+{
+  MFDClass::updateTopic();
+
+  std::scoped_lock<std::mutex> lock(mutex_);
+  const float prob_threshold =
+      std::min(1.0f, std::max(0.0f, scalar_threshold_property_->getFloat()));
+  log_odds_threshold_ = Bonxai::ProbabilisticMap::logods(prob_threshold);
 }
 
 void BonxaiGridDisplay::updateStyle()
@@ -125,7 +214,7 @@ void BonxaiGridDisplay::update([[maybe_unused]] float wall_dt,
     }
     else
     {
-      cloud_->setDimensions(voxel_size_, voxel_size_, voxel_size_);
+      cloud_->setDimensions(box_size_, box_size_, box_size_);
     }
 
     cloud_->addPoints(new_points_.begin(), new_points_.end());
@@ -164,6 +253,116 @@ bool BonxaiGridDisplay::updateFromTF()
   return true;
 }
 
+template <>
+bool ScalarBonxaiGridDisplay<float>::checkType(const std::string& type_id)
+{
+  return (type_id == "float");
+}
+
+bool ColorBonxaiGridDisplay::checkType(const std::string& type_id)
+{
+  // == does not work b/c ros type generation adds some fancy endings to
+  // the datatype
+  return (type_id.find("std_msgs::msg::ColorRGBA") != std::string::npos);
+}
+
+bool ProbabilisticBonxaiGridDisplay::checkType(const std::string& type_id)
+{
+  return (type_id == "Bonxai::ProbabilisticMap::CellT");
+}
+
+template <>
+bool ScalarBonxaiGridDisplay<float>::shouldShowCell(const float& cell)
+{
+  return (cell > scalar_threshold_property_->getFloat());
+}
+
+bool ProbabilisticBonxaiGridDisplay::shouldShowCell(const ProbabilisticCell& cell)
+{
+  return (cell.probability_log > log_odds_threshold_);
+}
+
+template <>
+void ScalarBonxaiGridDisplay<float>::setVoxelColor(
+    rviz_rendering::PointCloud::Point& new_point,
+    float& cell)
+{
+  BonxaiVoxelColorMode bonxai_color_mode =
+      static_cast<BonxaiVoxelColorMode>(bonxai_coloring_property_->getOptionInt());
+  float cell_probability{};
+  switch (bonxai_color_mode)
+  {
+    case BONXAI_CELL_COLOR:
+      setStatusStd(StatusProperty::Error,
+                   "Messages",
+                   "Cannot extract color from Scalar grid");
+      break;
+    case BONXAI_Z_AXIS_COLOR:
+      setColorFromZAxis(new_point.position.z, color_factor_, new_point);
+      break;
+    case BONXAI_PROBABILITY_COLOR:
+      // TODO (jjd9): Add min/max parameter for scaling between 0 and 1.
+      cell_probability = std::max(std::min(1.0f, cell), 0.0f);
+      // high probability -> green, low probability -> red
+      new_point.setColor((1.0f - cell_probability), cell_probability, 0.0);
+      break;
+    default:
+      break;
+  }
+}
+
+void ProbabilisticBonxaiGridDisplay::setVoxelColor(
+    rviz_rendering::PointCloud::Point& new_point,
+    ProbabilisticCell& cell)
+{
+  BonxaiVoxelColorMode bonxai_color_mode =
+      static_cast<BonxaiVoxelColorMode>(bonxai_coloring_property_->getOptionInt());
+  float cell_probability{};
+  switch (bonxai_color_mode)
+  {
+    case BONXAI_CELL_COLOR:
+      setStatusStd(StatusProperty::Error,
+                   "Messages",
+                   "Cannot extract color from Scalar grid");
+      break;
+    case BONXAI_Z_AXIS_COLOR:
+      setColorFromZAxis(new_point.position.z, color_factor_, new_point);
+      break;
+    case BONXAI_PROBABILITY_COLOR:
+      cell_probability = Bonxai::ProbabilisticMap::prob(cell.probability_log);
+      cell_probability = std::max(0.0f, std::min(cell_probability, 1.0f));
+      new_point.setColor((1.0f - cell_probability), cell_probability, 0.0);
+      break;
+    default:
+      break;
+  }
+}
+
+void ColorBonxaiGridDisplay::setVoxelColor(
+    rviz_rendering::PointCloud::Point& new_point,
+    std_msgs::msg::ColorRGBA& cell)
+{
+  BonxaiVoxelColorMode bonxai_color_mode =
+      static_cast<BonxaiVoxelColorMode>(bonxai_coloring_property_->getOptionInt());
+
+  switch (bonxai_color_mode)
+  {
+    case BONXAI_CELL_COLOR:
+      new_point.setColor(cell.r, cell.g, cell.b);
+      break;
+    case BONXAI_Z_AXIS_COLOR:
+      setColorFromZAxis(new_point.position.z, color_factor_, new_point);
+      break;
+    case BONXAI_PROBABILITY_COLOR:
+      setStatusStd(StatusProperty::Error,
+                   "Messages",
+                   "Cannot extract probability from Color grid.");
+      break;
+    default:
+      break;
+  }
+}
+
 template <typename CellT>
 void TemplatedBonxaiGridDisplay<CellT>::processMessage(
     const bonxai_msgs::msg::Bonxai::ConstSharedPtr msg)
@@ -187,7 +386,7 @@ void TemplatedBonxaiGridDisplay<CellT>::processMessage(
     // creating bonxai grid
     std::string grid_type;  // this will be CellT as a string
     Bonxai::VoxelGrid<CellT> grid = bonxai_msgs::fromRosMsg<CellT>(*msg, grid_type);
-    voxel_size_ = grid.resolution;
+    box_size_ = grid.resolution;
 
     setStatusStd(StatusProperty::Ok, "Type", grid_type.c_str());
     if (!checkType(grid_type))
@@ -233,7 +432,7 @@ void TemplatedBonxaiGridDisplay<CellT>::processMessage(
         return;
       }
 
-      const auto point = CoordToPos(coord, voxel_size_);
+      const auto point = CoordToPos(coord, box_size_);
       rviz_rendering::PointCloud::Point new_point;
       new_point.position.x = point.x;
       new_point.position.y = point.y;
@@ -271,6 +470,17 @@ void TemplatedBonxaiGridDisplay<CellT>::processMessage(
   }
 }
 
+using FloatBonxaiGridDisplay = ScalarBonxaiGridDisplay<float>;
 }  // namespace bonxai_rviz_plugins
 
 #include "pluginlib/class_list_macros.hpp"
+
+// TODO (jjd9): These plugins are SUPER specific (e.g. float and
+// std_msgs::msg::ColorRGBA, whereas bonxai supports all scalar values and all
+// rgb-like structs). It would be nice to make this more flexible somehow.
+PLUGINLIB_EXPORT_CLASS(bonxai_rviz_plugins::FloatBonxaiGridDisplay,
+                       rviz_common::Display)
+PLUGINLIB_EXPORT_CLASS(bonxai_rviz_plugins::ColorBonxaiGridDisplay,
+                       rviz_common::Display)
+PLUGINLIB_EXPORT_CLASS(bonxai_rviz_plugins::ProbabilisticBonxaiGridDisplay,
+                       rviz_common::Display)
